+++ src/SimpleTime.cs
@@ -18,16 +18,12 @@ namespace CampaignPacer
 		[SaveableProperty(4)]
 		public float FractionalDay { get; set; }
 
-		[SaveableProperty(5)]
-		public int DaysPerSeason { get; set; }
-
 		public SimpleTime()
 		{
 			Year = -1;
 			Season = -1;
 			Day = -1;
 			FractionalDay = -1f;
-			DaysPerSeason = -1;
 		}
 
 		public SimpleTime(CampaignTime ct)
@@ -44,14 +40,13 @@ namespace CampaignPacer
 			fracDays -= Day;
 
 			FractionalDay = MathF.Clamp((float)fracDays, 0f, 0.99999f);
-			DaysPerSeason = (int)Main.TimeParam.DayPerSeasonL;
 		}
 
 		public CampaignTime ToCampaignTime() => CampaignTime.Years(Year) + CampaignTime.Seasons(Season) + CampaignTime.Days(Day + FractionalDay);
 
-		public bool IsNull => Year == -1 && Season == -1 && Day == -1 && FractionalDay < -0.9 && DaysPerSeason == -1;
+		public bool IsNull => Year == -1 && Season == -1 && Day == -1 && FractionalDay < -0.99 && FractionalDay > -1.01;
 		public bool IsValid => IsNull || (Year >= 0 && IsSeasonValid && Day >= 0 && IsFractionalDayValid);
-		public bool IsValidWithCurrentCalendar => IsValid && Day < Main.TimeParam.DayPerSeasonL && DaysPerSeason == (int)Main.TimeParam.DayPerSeasonL;
+		public bool IsValidWithCurrentCalendar => IsValid && Day < Main.TimeParam.DayPerSeasonL;
 		private bool IsFractionalDayValid => FractionalDay >= 0f && FractionalDay < 1f;
 		private bool IsSeasonValid => Season >= 0 && Season < TimeParams.SeasonPerYearL;